Make integration tests less flaky. (#475)
Most test failures are happening because the host input box is closed
when quickstart page appears. This waits for the quickstart page to
appear before starting to run tests.

Renamed quickstart page so that it is unique in case some other
extension dependency (like python) decides to publish a quickstart page
and break our tests again.