Use structured output from sync command
This removes the need for complex state tracking and monitors for "start" and "complete"
events in the sync command stdout.

As we no longer forward log messages to the VS code terminal pane this change includes
rudimentary formatting of these messages into something human readable.

We should consider if it is worth keeping this pane or if the status should be
consolidated into the "sync" pane.