Add Pull Request Template (#603)
## Changes
<!-- Summary of your changes that are easy to understand -->
PR template to add structure for all pull requests. 

## Tests
<!-- How is this tested? -->
Copy pasting the template in this description to see how it looks, would
be able to test the changes in Github after this merges (for PRs after
this)

## Public Tickets
<!--
Link to public tickets (if any). Please don't mention internal tickets.
-->
N/A

---------

Co-authored-by: Fabian Jakobs <fabian.jakobs@databricks.com>