Fix multiple quote issue for .databricks.env file. (#686)
## Changes
* Variables being added to .databricks.env were being added with
multiple quotes. This was because of the following sequence of events:
* Vscode would try to search for user env file (.env) and not find it.
* It would then try to search for mspython python.envFile and not find
it.
* Finally it will set the userEnvFile to .env and the mspython
python.envFile to .databricks/.databricks.env.
* At the same time, if another instance of vscode is running in the same
project, it will go through the same steps, but this time find mspython
python.envFile to be .databricks/.databricks.env. It will use this value
for userEnvFile.
* Now both values are set to .databricks/.databricks.env. This is an
issue.

This is fixed by
* Stripping quotes
* More strict checking for .databricks/.databricks.env appearing in
userEnvFile.

This pr also adds an integration test for running .ipynb as jobs.

---------

Signed-off-by: kartikgupta-db <kartik.gupta@databricks.com>