Gracefully handle virtual /Repos/me folder (#693)
## Changes
The extension tries to find the root folder (/Repos/me or
/Users/me/.ide) depending on what the sync destination type is (repos or
workspace). The current root is stored in `currentFsRoot`. In case it
fails in finding the `currentFsRoot` (/Repos/me and /Users/me can be
virtual when a user hasn't interacted with these folders yet), it always
tries to create it under /User/me. This can be an issue if
`currentFsRoot` points to /Repos/me.

* This PR fixes this behaviour, by only allowing creation of
`workspaceFsRoot` (which always points to /Users/me/.ide) under /User/me
and not `currentFsRoot`.
* Also fix error reporting to show more verbose errors.

Creating a repo from the extension automatically handles generation of
/Repos/me folder as well.

Fixes #683  #691 #688 

## Tests
* Manually verified to work after the change.

---------

Co-authored-by: Fabian Jakobs <fabian.jakobs@databricks.com>