Show error for a corrupted config file (#955)
Better error messaging when config file is corrupted. Should help with
cases like this:
https://github.com/databricks/databricks-vscode/issues/942

When we detect profile loading/parsing error (that comes from the CLI),
we proceed with the configuration wizard without any pre-defined
profiles, but also show a warning notification to the users, prompting
them to open their config file and fix it.