Refactor config loaders (#978)
## Changes
This changes the `*ConfigReaderWriter` classes into `*ConfigModel`
classes. We introduce the following architecture:
* Configs storage is handled by "Storage Medium" such as `BundleFileSet`
and `StateStorage`. They emit event if there is a change. If we can't
monitor the underlying medium for changes, we must emit event from the
"set" method (see `StateStorage`).
* Models are responsible for extracting their respective domain of
configs from whatever storage medium they choose. They are also
responsible for emitting and event if there is a change in any of the
configs in their domain.
* Models might need some information such as target and auth params, to
automatically refresh the configs in their domain. There are multiple
setters which are exposed by each loader for taking this information. We
want to have multiple setters and not a single one because multiple
pieces of information can come at different times. Eg target -> load
auth info and login -> pass auth params to loader
* Models are responsible for correctly writing data to the storage
medium.

## Tests
<!-- How is this tested? -->