Add bundle project manager
"Bundle project manager" is being initialised before configModel or
connectionManager.
If necessary, it does automatic migration from the legacy projects to the
bundles (will be implemented in the follow-up).
If it detects multiple bundle projects and offers to choose which one to open.
If it detects a bundle config in the root of the workspace it kicks off
configModel and connectionManager.
If it detects no bundle config in the root, it offers to create a
project (bundle-init logic in the terminal will be in the follow-up)