Add bundle project manager (#1011)
"Bundle project manager" is being initialised before configModel or
connectionManager.
If necessary, it does automatic migration from the legacy projects to
the bundles (will be implemented in the follow-up).
If it detects multiple bundle projects and offers to choose which one to
open. If it detects a bundle config in the root of the workspace it
kicks off configModel and connectionManager.
If it detects no bundle config in the root, it offers to create a
project (spawning the Databricks CLI in the terminal)

I've played around with displaying notifications for things like "hey,
chose one project to open", or "there's no config in your workspace,
would you like to init a new project?", but they don't flow well with
existing quick-pick/input wizards and are pretty easy to miss. Currently
I'm using welcome views to show such information:

The default view is still the same:
![Screenshot 2024-01-22 at 10 11
20](https://github.com/databricks/databricks-vscode/assets/148094031/b185b327-5e91-4fed-8609-93724c9532c8)

If we don't detect bundle config in the root, but there are subfolders
with bundles, we show an additional text and a button:
![Screenshot 2024-01-22 at 10 11
58](https://github.com/databricks/databricks-vscode/assets/148094031/e90184ea-3e74-44b1-b0ba-f90720c7c72f)

It's still possible to login in this state, after which we will also
show "init" button:

![Screenshot 2024-01-22 at 10 13
27](https://github.com/databricks/databricks-vscode/assets/148094031/9b43a3b6-ac62-47cc-93db-1dd1285023c2)

If there are no sub-projects, then we just show login button alone in
the view:
![Screenshot 2024-01-22 at 10 14
54](https://github.com/databricks/databricks-vscode/assets/148094031/68f4a84f-4974-4ba3-9d96-45f7113668c2)