Improve login flow, properly init cluster manager
ConnectionManager first tries to login, and only after that starts listening to
configModel changes.

During login flow it updates configModel with authProvider, triggering reloading
of the config state. That's when we load clusterId from the bundle-validate model.

But the clusterId changes never trigger updateClusterManager, since they
happen before ConnectionManager is done with the login and before it started
listening to the changes.

Moving the manual update* calls after we set the auth provider (and awating till
the config model loads all the sub-configs data) fixes the problem.