Improve UX for bundle init folder selection (#1026)
Tried using showOpenDialog directly to select parent folder, but at
least on macOS we can't provide a context for that window (e.g. a title
explaining what are you selecting a folder for). Instead we are using
quick pick where you can write a path to a folder manually or chose to
click on a button that opens system's folder-selection dialog.