Prevent running resources when cli already launched but no run status (#1072)
## Changes
* When we launch a run, there is a lag between the CLI process execution
and when the IDE is able to poll for the first status update from the
API. We should not allow a second run of the resource during this time.
This PR blocks second runs by considering the "unknown" status also as a
running state for the viewitem context. This allows vscode to hide the
run button, even when we, we do not have a status update. Errors and
cancellations are still handled gracefully, because the status reaches
completed state.

* Also, clear the run statuses map on target change. 
## Tests
<!-- How is this tested? -->