Fix env loading in notebooks
In monorepos it's possible to have multiple `.databricks` folders, while
our extension will only put .env file in the root of the workspace.

In that case notebook bootstrap code was trying (and failing) to loading
`.env` file from the first `.databricks` folder.

This change makes sure we load the `.env` file from the first folder
that actually has it.