Bump CLI to 228.1 and generate new types
CLI changed it schema output, and now every field can also be a string
(due to variables). This breaks many usages of the schema types, since we
don't expect strings to be everywhere, as we rely not on the raw yaml config,
but on the 'bundle validate' (or 'summary') output, which extraploates variables.

Had to resort to typescript magic to strip down strings from all union types.