Add partial update to pipelines
The icon is WIP, I'll contact designers to get a better one.
Currently we don't pre-load pipeline updates, so if user hasn't
executed a pipeline from the extension, the partial update action
will not show a list of tables to select from (and users will have
to type the table name manually). After the pipeline is executed we
show the tables selection in the partial update action.

Pre-loading pipeline updates is for later.