Configure autocomplete as a part of env panel
"Configure autocomplete" logic was disconnected from the enc setup,
and it was still possible to install pyspark into the global python installation.

Or re-install it, even if we already installed it before through env panel.

Now "Setup Databricks builtins" appears in the env panel as an optional step
if all the previous env checks are green.

Also fix a bug where we were settig up builtins file in the bundle root,
while the python extension expects a workspace root.