Fix %run magic with databricks notebooks on windows (#1628)
## Changes

On windows `tempfile.NamedTemporaryFile` with `delete=True` (default)
deletes the temp file when it's closed. Internally the `%run` logic
opens and closes the file before execution (to check that it can open
it), which leads to deletion of the file and an error seen in #1624

Here we use lower-level temp file logic and unlink it ourselves after
the execution is fully done.

One more change is an installation of `nbformat` dependency at runtime
(together with `databricks-connect`), which is required for the `%run`
to work.

## Tests

e2e